Fix: ReceiveMaximum is a 2 byte integer

+++ src/packets/connect.rs
@@ -12,7 +12,7 @@ crate::properties::define_properties! {
         session_expiry_interval: SessionExpiryInterval with setter = u32,
 
         (anker: "_Toc3901049")
-        receive_maximum: ReceiveMaximum with setter = u32,
+        receive_maximum: ReceiveMaximum with setter = u16,
 
         (anker: "_Toc3901050")
         maximum_packet_size: MaximumPacketSize with setter = u32,
